Completed the module
